improve data flow
